Gupo #1

+++ Inspinia_MVC5_SeedProject/Models/UDV_Vent_Busqueda_Factura.cs
@@ -21,7 +21,7 @@ namespace ERP_GMEDINA.Models
         public string esfac_Descripcion { get; set; }
         public short cja_Id { get; set; }
         public string cja_Descripcion { get; set; }
-        public short suc_Id { get; set; }
+        public int suc_Id { get; set; }
         public int clte_Id { get; set; }
         public string pemi_NumeroCAI { get; set; }
         public bool fact_AlCredito { get; set; }
